refactor: consolidate user metadata lookup logic
- Removed the PrepareForMetadataLookup method from the User model and replaced it with a MetadataLookup method in the userReaderWriter struct.
- Updated the GetUserMetadata method to utilize the new MetadataLookup method for improved clarity and functionality.
- Added a new Auth0User model to represent users in Auth0, including conversion methods to the domain User model.

Jira Ticket: https://linuxfoundation.atlassian.net/browse/LFXV2-493
