+++ internal/domain/model/user.go
@@ -237,22 +237,3 @@ func (a *UserMetadata) Patch(update *UserMetadata) bool {
 
 	return updated
 }
-
-// PrepareForMetadataLookup prepares the user for metadata lookup based on the input
-// Returns true if should use canonical lookup, false if should use search
-func (u *User) PrepareForMetadataLookup(input string) bool {
-	input = strings.TrimSpace(input)
-
-	if strings.Contains(input, "|") {
-		// Input contains "|", use as sub for canonical lookup
-		u.Sub = input
-		u.UserID = input // Auth0 uses user_id for the canonical lookup
-		u.Username = ""  // Clear username field
-		return true
-	}
-	// Input doesn't contain "|", use for search query
-	u.Sub = ""    // Clear sub field
-	u.UserID = "" // Clear user_id field
-	u.Username = input
-	return false
-}
