+++ internal/domain/model/user_test.go
@@ -394,229 +394,6 @@ func TestUserMetadata_userMetadataSanitize(t *testing.T) {
 	})
 }
 
-func TestUser_PrepareForMetadataLookup(t *testing.T) {
-	tests := []struct {
-		name              string
-		input             string
-		expectedCanonical bool
-		expectedSub       string
-		expectedUserID    string
-		expectedUsername  string
-	}{
-		{
-			name:              "canonical lookup with auth0 connection",
-			input:             "auth0|123456789",
-			expectedCanonical: true,
-			expectedSub:       "auth0|123456789",
-			expectedUserID:    "auth0|123456789",
-			expectedUsername:  "",
-		},
-		{
-			name:              "canonical lookup with google oauth2",
-			input:             "google-oauth2|987654321",
-			expectedCanonical: true,
-			expectedSub:       "google-oauth2|987654321",
-			expectedUserID:    "google-oauth2|987654321",
-			expectedUsername:  "",
-		},
-		{
-			name:              "canonical lookup with github connection",
-			input:             "github|456789123",
-			expectedCanonical: true,
-			expectedSub:       "github|456789123",
-			expectedUserID:    "github|456789123",
-			expectedUsername:  "",
-		},
-		{
-			name:              "canonical lookup with samlp enterprise",
-			input:             "samlp|enterprise|user123",
-			expectedCanonical: true,
-			expectedSub:       "samlp|enterprise|user123",
-			expectedUserID:    "samlp|enterprise|user123",
-			expectedUsername:  "",
-		},
-		{
-			name:              "canonical lookup with linkedin",
-			input:             "linkedin|789123456",
-			expectedCanonical: true,
-			expectedSub:       "linkedin|789123456",
-			expectedUserID:    "linkedin|789123456",
-			expectedUsername:  "",
-		},
-		{
-			name:              "search lookup with simple username",
-			input:             "john.doe",
-			expectedCanonical: false,
-			expectedSub:       "",
-			expectedUserID:    "",
-			expectedUsername:  "john.doe",
-		},
-		{
-			name:              "search lookup with complex username",
-			input:             "jane.smith123",
-			expectedCanonical: false,
-			expectedSub:       "",
-			expectedUserID:    "",
-			expectedUsername:  "jane.smith123",
-		},
-		{
-			name:              "search lookup with developer username",
-			input:             "developer123",
-			expectedCanonical: false,
-			expectedSub:       "",
-			expectedUserID:    "",
-			expectedUsername:  "developer123",
-		},
-		{
-			name:              "canonical lookup with spaces (trimmed)",
-			input:             "  auth0|123456789  ",
-			expectedCanonical: true,
-			expectedSub:       "auth0|123456789",
-			expectedUserID:    "auth0|123456789",
-			expectedUsername:  "",
-		},
-		{
-			name:              "search lookup with spaces (trimmed)",
-			input:             "  john.doe  ",
-			expectedCanonical: false,
-			expectedSub:       "",
-			expectedUserID:    "",
-			expectedUsername:  "john.doe",
-		},
-		{
-			name:              "canonical lookup with pipe at start",
-			input:             "|startpipe",
-			expectedCanonical: true,
-			expectedSub:       "|startpipe",
-			expectedUserID:    "|startpipe",
-			expectedUsername:  "",
-		},
-		{
-			name:              "canonical lookup with pipe at end",
-			input:             "endpipe|",
-			expectedCanonical: true,
-			expectedSub:       "endpipe|",
-			expectedUserID:    "endpipe|",
-			expectedUsername:  "",
-		},
-		{
-			name:              "empty input (search strategy)",
-			input:             "",
-			expectedCanonical: false,
-			expectedSub:       "",
-			expectedUserID:    "",
-			expectedUsername:  "",
-		},
-	}
-
-	for _, tt := range tests {
-		t.Run(tt.name, func(t *testing.T) {
-			user := &User{}
-			result := user.PrepareForMetadataLookup(tt.input)
-
-			// Check strategy result
-			if result != tt.expectedCanonical {
-				t.Errorf("PrepareForMetadataLookup() returned %v, expected %v", result, tt.expectedCanonical)
-			}
-
-			// Check Sub field
-			if user.Sub != tt.expectedSub {
-				t.Errorf("Sub = %q, expected %q", user.Sub, tt.expectedSub)
-			}
-
-			// Check UserID field
-			if user.UserID != tt.expectedUserID {
-				t.Errorf("UserID = %q, expected %q", user.UserID, tt.expectedUserID)
-			}
-
-			// Check Username field
-			if user.Username != tt.expectedUsername {
-				t.Errorf("Username = %q, expected %q", user.Username, tt.expectedUsername)
-			}
-		})
-	}
-}
-
-func TestUser_PrepareForMetadataLookup_Idempotency(t *testing.T) {
-	// Test that calling PrepareForMetadataLookup multiple times with the same input
-	// produces the same result
-	user := &User{}
-
-	// First call
-	result1 := user.PrepareForMetadataLookup("auth0|123456789")
-	sub1, userID1, username1 := user.Sub, user.UserID, user.Username
-
-	// Second call with same input
-	result2 := user.PrepareForMetadataLookup("auth0|123456789")
-	sub2, userID2, username2 := user.Sub, user.UserID, user.Username
-
-	// Results should be identical
-	if result1 != result2 {
-		t.Errorf("PrepareForMetadataLookup() not idempotent: first=%v, second=%v", result1, result2)
-	}
-	if sub1 != sub2 {
-		t.Errorf("Sub not idempotent: first=%q, second=%q", sub1, sub2)
-	}
-	if userID1 != userID2 {
-		t.Errorf("UserID not idempotent: first=%q, second=%q", userID1, userID2)
-	}
-	if username1 != username2 {
-		t.Errorf("Username not idempotent: first=%q, second=%q", username1, username2)
-	}
-}
-
-func TestUser_PrepareForMetadataLookup_StrategySwitch(t *testing.T) {
-	// Test that switching between strategies properly clears the other fields
-	user := &User{}
-
-	// Start with canonical lookup
-	canonical := user.PrepareForMetadataLookup("auth0|123456789")
-	if !canonical {
-		t.Fatal("Expected canonical strategy")
-	}
-	if user.Sub == "" || user.UserID == "" || user.Username != "" {
-		t.Errorf("Canonical strategy fields not set correctly: Sub=%q, UserID=%q, Username=%q",
-			user.Sub, user.UserID, user.Username)
-	}
-
-	// Switch to search lookup
-	search := user.PrepareForMetadataLookup("john.doe")
-	if search {
-		t.Fatal("Expected search strategy")
-	}
-	if user.Sub != "" || user.UserID != "" || user.Username == "" {
-		t.Errorf("Search strategy fields not set correctly: Sub=%q, UserID=%q, Username=%q",
-			user.Sub, user.UserID, user.Username)
-	}
-}
-
-func TestUser_PrepareForMetadataLookup_PreservesOtherFields(t *testing.T) {
-	// Test that PrepareForMetadataLookup doesn't modify other user fields
-	user := &User{
-		Token:        "test-token",
-		PrimaryEmail: "test@example.com",
-		UserMetadata: &UserMetadata{
-			Name: converters.StringPtr("Test User"),
-		},
-	}
-
-	originalToken := user.Token
-	originalEmail := user.PrimaryEmail
-	originalMetadata := user.UserMetadata
-
-	user.PrepareForMetadataLookup("auth0|123456789")
-
-	if user.Token != originalToken {
-		t.Errorf("Token was modified: got %q, expected %q", user.Token, originalToken)
-	}
-	if user.PrimaryEmail != originalEmail {
-		t.Errorf("PrimaryEmail was modified: got %q, expected %q", user.PrimaryEmail, originalEmail)
-	}
-	if user.UserMetadata != originalMetadata {
-		t.Errorf("UserMetadata was modified")
-	}
-}
-
 func TestUser_buildIndexKey(t *testing.T) {
 	tests := []struct {
 		name         string
