+++ internal/domain/port/user.go
@@ -19,6 +19,7 @@ type UserReaderWriter interface {
 type UserReader interface {
 	GetUser(ctx context.Context, user *model.User) (*model.User, error)
 	SearchUser(ctx context.Context, user *model.User, criteria string) (*model.User, error)
+	MetadataLookup(ctx context.Context, input string, user *model.User) bool
 }
 
 // UserWriter defines the behavior of the user writer
